spark/requirements.txt
pyspark==3.4.1
kafka-python==2.0.2
pyiceberg[s3fs,hive]==0.5.1
boto3==1.28.62
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
